+++ src/QuickMarcEditor/QuickMarcEditorRows/QuickMarcEditorRows.test.js
@@ -67,7 +67,7 @@ const initValues = [
   {
     id: '9',
     tag: '100',
-    content: '$a c',
+    content: '$a Kirby, Jack, $e creator. $0 http://id.loc.gov/authorities/names/n2019022493',
     authorityId: 'authority-id',
     _isLinked: true,
     indicators: [],
@@ -314,6 +314,10 @@ describe('Given QuickMarcEditorRows', () => {
       fireEvent.click(getAllByTestId('link-authority-button')[0]);
       rerender(getComponent(props));
       expect(getByTestId('view-authority-record-link')).toBeVisible();
+      expect(getByTestId('view-authority-record-link')).toHaveAttribute(
+        'href',
+        '/marc-authorities/authorities/authority-id?authRefType=Authorized&headingRef=Kirby, Jack,&segment=search',
+      );
     });
   });
 });